requirements.txt
# Database
psycopg2-binary
requests
python-dotenv
SQLAlchemy

#object detection
#ultralytics
Flask